Add files via upload
First version of the bash to pre-processing of the OpenFoam simulation. It opens a screen asking for the main inputs for the simulation: Domain size, amount of nodes in x and y, amount of tumors, if with circular or elliptical shape, rotational angle to each tumor and it plots the position in a screen to pre-visualization.

Obs.: In the new version i´ll add a python code inside the bash to change the openfoam files with libraries in it.